apps/multimodal-processor/requirements.txt
# Multimodal Processor Dependencies

# FastAPI and Server
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Image Processing & OCR
pillow==10.2.0
pytesseract==0.3.10
opencv-python-headless==4.9.0.80

# PDF Processing
pymupdf==1.23.26

# HTTP Client
httpx==0.26.0

# Data Validation
pydantic==2.6.1
pydantic-settings==2.1.0

# Logging
python-json-logger==2.0.7

# Numpy
numpy==1.26.4
